fix sign conversion bug (int -> size_t)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@1241 3db68df2-c116-0410-a063-a993310a9797

+++ source/gui/CInput.cpp
@@ -287,7 +287,7 @@ void CInput::Draw()
 		// U+FE33: PRESENTATION FORM FOR VERTICAL LOW LINE
 		// (sort of like a | which is aligned to the left of most characters)
 
-		for (int i=0; i<caption.Length(); ++i)
+		for (size_t i=0; i<caption.Length(); ++i)
 		{
 			if (m_iBufferPos == i)
 			{
